+++ cocos/physics/framework/components/colliders/mesh-collider.ts
@@ -11,7 +11,7 @@ import {
     editable,
     serializable,
 } from 'cc.decorator';
-import { Collider } from './collider-component';
+import { Collider } from './collider';
 import { Mesh } from '../../../../core';
 import { ITrimeshShape } from '../../../spec/i-physics-shape';
 import { EDITOR, TEST } from 'internal:constants';
