+++ cocos/physics/framework/components/colliders/plane-collider.ts
@@ -13,7 +13,7 @@ import {
     serializable,
 } from 'cc.decorator';
 import { Vec3 } from '../../../../core/math';
-import { Collider } from './collider-component';
+import { Collider } from './collider';
 import { IPlaneShape } from '../../../spec/i-physics-shape';
 import { EDITOR, TEST } from 'internal:constants';
 import { EColliderType } from '../../physics-enum';
