+++ cocos/physics/framework/components/colliders/simplex-collider.ts
@@ -13,7 +13,7 @@ import {
     serializable,
 } from 'cc.decorator';
 import { Vec3 } from '../../../../core/math';
-import { Collider } from './collider-component';
+import { Collider } from './collider';
 import { ISimplexShape } from '../../../spec/i-physics-shape';
 import { EDITOR, TEST } from 'internal:constants';
 import { ESimplexType, EColliderType } from '../../physics-enum';
