+++ cocos/physics/framework/components/constant-force.ts
@@ -14,7 +14,7 @@ import {
     serializable,
 } from 'cc.decorator';
 import { Component } from '../../../core/components/component';
-import { RigidBody } from './rigid-body-component';
+import { RigidBody } from './rigid-body';
 import { Vec3 } from '../../../core/math/vec3';
 import { EDITOR } from 'internal:constants';
